split libs.h into multiple files
